requirements.txt
pandas==1.2.5
requests==2.25.1
streamlit==0.88.0
Pillow==8.3.2
